Rearrange source repo
