+++ src/ghosthunter-nodependency.js
